fix: Complete AILib integration after directory reorganization
- Fix all include paths in ailib source files to use new structure
- Update legacy CQL files to reference moved headers (secure_store.hpp, json_utils.hpp)
- Add missing secure_store.cpp to CMakeLists.txt for SecureString linking
- Fix ailib test files to use correct include paths
- Main cql executable now builds successfully with new AILib structure

Integration changes:
- lib/ailib/src files now use ailib/ include paths
- src/cql files updated to reference moved ailib headers
- CMakeLists.txt includes all necessary ailib source files
- Test files updated for new structure
